Rename editing views
So as to separate them from non-editing ones
