+++ editorsnotes_app/src/views/edit_document.js
