+++ editorsnotes_app/src/views/edit_note.js
@@ -4,7 +4,7 @@ var Backbone = require('../backbone')
   , Cocktail = require('backbone.cocktail')
   , _ = require('underscore')
   , i18n = require('../utils/i18n').main
-  , NoteSectionListView = require('./note_section_list')
+  , NoteSectionListView = require('./edit_note_section_list')
   , RelatedTopicsView = require('./related_topics')
   , SaveItemMixin = require('./generic/save_item_mixin')
   , HandleErrorMixin = require('./generic/handle_error_mixin')
