+++ editorsnotes_app/src/views/edit_note_section.js
