+++ editorsnotes_app/src/views/edit_topic_citations.js
@@ -2,7 +2,26 @@
 
 var OrderedCollectionView = require('./generic/ordered_collection_base')
   , $ = require('jquery')
-  , CitationView = require('./citation')
+  , CitationSectionView = require('./note_section').citation
+  , CitationView
+
+CitationView = CitationSectionView.extend({
+  initialize: function () { this.render(); },
+  render: function () {
+    var template = require('../templates/note_section_citation.html')
+      , citation = this.model
+      , html
+
+    html = template({ns: {
+      'document': citation.get('document'),
+      'document_description': citation.get('document_description'),
+      'content': citation.get('notes')
+    }});
+
+    this.$el.html(html);
+    this.afterRender();
+  }
+});
 
 module.exports = OrderedCollectionView.extend({
   itemViewConstructor: CitationView,
